Ønsker å kunne gjenbruke hooket for å hente informasjon om nedlastbar fil, har derfor tatt koden ut i hooks
